pyproject.toml
[project]
name = "alzheimer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google>=3.0.0",
    "google-generativeai>=0.8.6",
    "networkx>=3.6.1",
    "python-dotenv>=1.2.2",
]
